pyproject.toml
[tool.poetry]
name = "lsm-system-drf"
version = "0.1.0"
description = ""
authors = ["DalOpez <alex-grazhdankin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.3"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.0"
pillow = "^11.2.1"
black = "^25.1.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
